+++ src/main/scala/dev/mongocamp/driver/mongodb/pagination/MongoPagination.scala
@@ -3,7 +3,7 @@ package dev.mongocamp.driver.mongodb.pagination
 import dev.mongocamp.driver.mongodb.database.ConfigHelper
 
 trait MongoPagination[A <: Any] extends ConfigHelper {
-  def paginate(page: Long, rows: Long): PaginationResult[A]
+  def paginate(page: Int, rows: Int): PaginationResult[A]
   def countResult: Long
 
   def foreach(a: A => Unit): Unit = {
